+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+12 Jan
+  - Skip move command to generate moves to switch playing side in analysis
+
 27 Dec
   - Learn from your mistakes in Studies
   - Pin study on home page